Modified trees to accept doubles and added flower set

+++ src/tree/tree.cpp
@@ -1,9 +1,9 @@
 #include "tree.hpp"
 #include <set>
 
-Node *Tree::_build_tree(std::vector<std::vector<int>> rows)
+Node *Tree::_build_tree(std::vector<std::vector<double>> rows)
 {
-	std::tuple<int, Question *> gain_and_question = find_best_split(rows);
+	std::tuple<double, Question *> gain_and_question = find_best_split(rows);
 	int gain = std::get<0>(gain_and_question);
 	Question *question = std::get<1>(gain_and_question);
 
@@ -12,13 +12,13 @@ Node *Tree::_build_tree(std::vector<std::vector<int>> rows)
 		return new EndNode(rows);
 	}
 
-	std::tuple<std::vector<std::vector<int>>,
-			   std::vector<std::vector<int>>>
+	std::tuple<std::vector<std::vector<double>>,
+			   std::vector<std::vector<double>>>
 		true_and_false_rows = partition(rows, question);
 
-	std::vector<std::vector<int>> true_rows =
+	std::vector<std::vector<double>> true_rows =
 		std::get<0>(true_and_false_rows);
-	std::vector<std::vector<int>> false_rows =
+	std::vector<std::vector<double>> false_rows =
 		std::get<1>(true_and_false_rows);
 
 	Node *true_branch = _build_tree(true_rows);
@@ -27,7 +27,7 @@ Node *Tree::_build_tree(std::vector<std::vector<int>> rows)
 	return new DecisionNode(question, true_branch, false_branch);
 }
 
-std::map<int, int> class_counts(std::vector<std::vector<int>> rows)
+std::map<int, int> class_counts(std::vector<std::vector<double>> rows)
 {
 	std::map<int, int> counts;
 	for (auto row : rows)
@@ -38,7 +38,7 @@ std::map<int, int> class_counts(std::vector<std::vector<int>> rows)
 	return counts;
 }
 
-double Tree::gini(std::vector<std::vector<int>> rows)
+double Tree::gini(std::vector<std::vector<double>> rows)
 {
 	std::map<int, int> counts = class_counts(rows);
 	double impurity = 1;
@@ -50,13 +50,13 @@ double Tree::gini(std::vector<std::vector<int>> rows)
 	return impurity;
 }
 
-std::tuple<std::vector<std::vector<int>>,
-		   std::vector<std::vector<int>>>
-Tree::partition(std::vector<std::vector<int>> rows,
+std::tuple<std::vector<std::vector<double>>,
+		   std::vector<std::vector<double>>>
+Tree::partition(std::vector<std::vector<double>> rows,
 				Question *question)
 {
-	std::vector<std::vector<int>> true_rows;
-	std::vector<std::vector<int>> false_rows;
+	std::vector<std::vector<double>> true_rows;
+	std::vector<std::vector<double>> false_rows;
 
 	for (auto row : rows)
 	{
@@ -73,15 +73,15 @@ Tree::partition(std::vector<std::vector<int>> rows,
 	return std::make_tuple(true_rows, false_rows);
 }
 
-double Tree::info_gain(std::vector<std::vector<int>> left,
-					   std::vector<std::vector<int>> right,
+double Tree::info_gain(std::vector<std::vector<double>> left,
+					   std::vector<std::vector<double>> right,
 					   double current_uncertainty)
 {
 	double p = left.size() / (left.size() + right.size());
 	return current_uncertainty - p * gini(left) - (1 - p) * gini(right);
 }
 
-std::tuple<int, Question *> Tree::find_best_split(std::vector<std::vector<int>> rows)
+std::tuple<double, Question *> Tree::find_best_split(std::vector<std::vector<double>> rows)
 {
 	int best_gain = 0;
 	Question *best_question;
@@ -90,7 +90,7 @@ std::tuple<int, Question *> Tree::find_best_split(std::vector<std::vector<int>> 
 
 	for (int column = 0; column < n_features; column++)
 	{
-		std::set<int> values;
+		std::set<double> values;
 		for (auto row : rows)
 		{
 			values.insert(row[column]);
@@ -98,13 +98,13 @@ std::tuple<int, Question *> Tree::find_best_split(std::vector<std::vector<int>> 
 		for (auto value : values)
 		{
 			Question *question = new Question(column, value);
-			std::tuple<std::vector<std::vector<int>>,
-					   std::vector<std::vector<int>>>
+			std::tuple<std::vector<std::vector<double>>,
+					   std::vector<std::vector<double>>>
 				true_and_false_rows = partition(rows, question);
 
-			std::vector<std::vector<int>> true_rows =
+			std::vector<std::vector<double>> true_rows =
 				std::get<0>(true_and_false_rows);
-			std::vector<std::vector<int>> false_rows =
+			std::vector<std::vector<double>> false_rows =
 				std::get<1>(true_and_false_rows);
 
 			if (true_rows.empty() || false_rows.empty())
@@ -135,7 +135,7 @@ const std::string Tree::to_string()
 	return tree_str;
 }
 
-EndNode *Tree::_classify(std::vector<int> row, Node *node)
+EndNode *Tree::_classify(std::vector<double> row, Node *node)
 {
 	if (node->get_type() == NodeType::ENDNODE)
 	{
